# ARROW-3922: [C++] Micro-optimizations to BitUtil::GetBit