@wesm 2018-12-01
The implementation of BitmapReader is unchanged so you can see the noise comparing the two benchmarks which supposedly should be the same