@animeshtrivedi 2018-12-02
Hi @wesm , thanks for trying this optimization out. Generally I would have expected that a bit of variation in the performance, nonetheless as you said since it generates less assembly, it is worth the change. 
At least in my microbenchmarks (reading 10 billion integers in memory), the bitmap change resulted in ~50% performance gains ;)