@wesm 2018-12-02
If you could contribute to our benchmark suite that would be awesome